+++ src/wm/host_func.rs
@@ -173,8 +173,7 @@ pub fn call(
                 owner,
                 caller: ctx.owner,
             };
-
-            let app_hash = wm.app_hash_check(ctx.db, ctx.owner, contract, ctx_args)?;
+            let app_hash = wm.app_hash_check(ctx.db,  contract, ctx_args)?;
             wm.call(
                 ctx.db,
                 ctx.depth + 1,